Prod 2019

+++ uber/api.py
@@ -16,8 +16,8 @@ from uber.barcode import get_badge_num_from_barcode
 from uber.config import c
 from uber.decorators import department_id_adapter
 from uber.errors import CSRFException
-from uber.models import AdminAccount, ApiToken, Attendee, Department, DeptMembership, DeptMembershipRequest, Job, \
-    Session, Shift, GuestGroup
+from uber.models import AdminAccount, ApiToken, Attendee, Department, DeptMembership, DeptMembershipRequest, \
+    IndieStudio, Job, Session, Shift, GuestGroup
 from uber.server import register_jsonrpc
 from uber.utils import check_csrf, normalize_newlines
 
@@ -99,7 +99,7 @@ def auth_by_token(required_access):
 def auth_by_session(required_access):
     try:
         check_csrf()
-    except CSRFException as ex:
+    except CSRFException:
         return (403, 'Your CSRF token is invalid. Please go back and try again.')
     admin_account_id = cherrypy.session.get('account_id', None)
     if not admin_account_id:
@@ -199,6 +199,50 @@ class GuestLookup:
 
 
 @all_api_auth(c.API_READ)
+class MivsLookup:
+    fields = {
+        'name': True,
+        'address': True,
+        'website': True,
+        'twitter': True,
+        'facebook': True,
+        'status_label': True,
+        'staff_notes': True,
+        'group': {
+            'name': True,
+        },
+        'developers': {
+            'full_name': True,
+            'first_name': True,
+            'last_name': True,
+            'email': True,
+            'cellphone': True,
+        },
+    }
+
+    def statuses(self):
+        return c.MIVS_STUDIO_STATUS_VARS
+
+    def list(self, status=False):
+        """
+        Returns a list of MIVS studios and their developers.
+
+        Optionally, 'status' may be passed to limit the results to MIVS
+        studios with a specific status. Use 'confirmed' to get MIVS teams
+        who are attending the event.
+
+        For a full list of statuses, call the "mivs.statuses" method.
+
+        """
+        with Session() as session:
+            if status and status.upper() in c.MIVS_STUDIO_STATUS_VARS:
+                query = session.query(IndieStudio).filter_by(status=getattr(c, status.upper()))
+            else:
+                query = session.query(IndieStudio)
+            return [mivs.to_dict(self.fields) for mivs in query]
+
+
+@all_api_auth(c.API_READ)
 class AttendeeLookup:
     fields = {
         'full_name': True,
@@ -240,7 +284,10 @@ class AttendeeLookup:
                 'type_label', 'department_name', 'name', 'description',
                 'start_time', 'end_time', 'extra15'
             ]
-        }
+        },
+        'group': {
+            'name': True,
+        },
     })
 
     def lookup(self, badge_num, full=False):
@@ -260,7 +307,7 @@ class AttendeeLookup:
             if attendee:
                 return attendee.to_dict(fields)
             else:
-                return {'error': 'No attendee found with Badge #{}'.format(badge_num)}
+                raise HTTPError(404, 'No attendee found with badge #{}'.format(badge_num))
 
     def search(self, query, full=False):
         """
@@ -281,28 +328,46 @@ class AttendeeLookup:
 
     def export(self, query, full=False):
         """
-        Searches for attendees by either email or first and last name.
+        Searches for attendees by either email, "first last" name, or
+        "first last &lt;email&gt;" combinations.
 
-        `query` should be a comma or newline separated list of emails and
-        "first last" name combos.
+        `query` should be a comma or newline separated list of email/name
+        queries.
+
+        Example:
+        <pre>Merrium Webster, only.email@example.com, John Doe &lt;jdoe@example.com&gt;</pre>
 
         Results are returned in the format expected by
         <a href="../import/staff">the staff importer</a>.
         """
-        queries = [s.strip() for s in re.split('[\n,]', normalize_newlines(query)) if s.strip()]
+        _re_name_email = re.compile(r'^\s*(.*?)\s*<\s*(.*?@.*?)\s*>\s*$')
+        _re_sep = re.compile(r'[\n,]')
+        _re_whitespace = re.compile(r'\s+')
+        queries = [s.strip() for s in _re_sep.split(normalize_newlines(query)) if s.strip()]
 
         names = dict()
         emails = dict()
+        names_and_emails = dict()
         ids = set()
         for q in queries:
             if '@' in q:
-                emails[Attendee.normalize_email(q)] = q
+                match = _re_name_email.match(q)
+                if match:
+                    name = match.group(1)
+                    email = Attendee.normalize_email(match.group(2))
+                    if name:
+                        first, last = (_re_whitespace.split(name.lower(), 1) + [''])[0:2]
+                        names_and_emails[(first, last, email)] = q
+                    else:
+                        emails[email] = q
+                else:
+                    emails[Attendee.normalize_email(q)] = q
             elif q:
                 try:
                     ids.add(str(uuid.UUID(q)))
                 except Exception:
-                    first, _, last = [s.strip() for s in q.partition(' ')]
-                    names[q] = (first.lower(), last.lower())
+                    first, last = (_re_whitespace.split(q.lower(), 1) + [''])[0:2]
+                    names[(first, last)] = q
 
         with Session() as session:
             if full:
@@ -318,27 +383,46 @@ class AttendeeLookup:
                     .options(*options).order_by(Attendee.email, Attendee.id).all()
 
             known_emails = set(a.normalized_email for a in email_attendees)
-            unknown_emails = sorted([email for normalized, email in emails.items() if normalized not in known_emails])
+            unknown_emails = sorted([raw for normalized, raw in emails.items() if normalized not in known_emails])
 
             name_attendees = []
             if names:
                 filters = [
-                    and_(func.lower(Attendee.first_name) == n[0], func.lower(Attendee.last_name) == n[1])
-                    for n in names.values()]
+                    and_(func.lower(Attendee.first_name) == first, func.lower(Attendee.last_name) == last)
+                    for first, last in names.keys()]
                 name_attendees = session.query(Attendee).filter(or_(*filters)) \
                     .options(*options).order_by(Attendee.email, Attendee.id).all()
+
+            known_names = set((a.first_name.lower(), a.last_name.lower()) for a in name_attendees)
+            unknown_names = sorted([raw for normalized, raw in names.items() if normalized not in known_names])
+
+            name_and_email_attendees = []
+            if names_and_emails:
+                filters = [
+                    and_(
+                        func.lower(Attendee.first_name) == first,
+                        func.lower(Attendee.last_name) == last,
+                        Attendee.normalized_email == email)
+                    for first, last, email in names_and_emails.keys()]
+                name_and_email_attendees = session.query(Attendee).filter(or_(*filters)) \
+                    .options(*options).order_by(Attendee.email, Attendee.id).all()
+
+            known_names_and_emails = set(
+                (a.first_name.lower(), a.last_name.lower(), a.normalized_email) for a in name_and_email_attendees)
+            unknown_names_and_emails = sorted([
+                raw for normalized, raw in names_and_emails.items() if normalized not in known_names_and_emails])
 
             id_attendees = []
             if ids:
                 id_attendees = session.query(Attendee).filter(Attendee.id.in_(ids)) \
                     .options(*options).order_by(Attendee.email, Attendee.id).all()
 
-            known_names = set(a.full_name.lower() for a in name_attendees)
-            unknown_names = sorted([full_name for full_name in names.keys() if full_name.lower() not in known_names])
+            known_ids = set(str(a.id) for a in id_attendees)
+            unknown_ids = sorted([i for i in ids if i not in known_ids])
 
             seen = set()
             all_attendees = [
-                a for a in (id_attendees + email_attendees + name_attendees)
+                a for a in (id_attendees + email_attendees + name_attendees + name_and_email_attendees)
                 if a.id not in seen and not seen.add(a.id)]
 
             fields = [
@@ -357,7 +441,12 @@ class AttendeeLookup:
                 'comments',
                 'admin_notes',
                 'all_years',
+                'badge_status',
+                'badge_status_label',
             ]
+            if full:
+                fields.extend(['shirt'])
+
             attendees = []
             for a in all_attendees:
                 d = a.to_dict(fields)
@@ -388,8 +477,10 @@ class AttendeeLookup:
                 attendees.append(d)
 
             return {
+                'unknown_ids': unknown_ids,
                 'unknown_emails': unknown_emails,
                 'unknown_names': unknown_names,
+                'unknown_names_and_emails': unknown_names_and_emails,
                 'attendees': attendees,
             }
 
@@ -458,7 +549,7 @@ class JobLookup:
         with Session() as session:
             message = session.assign(attendee_id, job_id)
             if message:
-                return {'error': message}
+                raise HTTPError(400, message)
             else:
                 session.commit()
                 return session.job(job_id).to_dict(self.fields)
@@ -470,14 +561,13 @@ class JobLookup:
         Takes the shift id as the only parameter.
         """
         with Session() as session:
-            try:
-                shift = session.shift(shift_id)
-                session.delete(shift)
-                session.commit()
-            except Exception:
-                return {'error': 'Shift was already deleted'}
-            else:
-                return session.job(shift.job_id).to_dict(self.fields)
+            shift = session.query(Shift).filter_by(id=shift_id).first()
+            if not shift:
+                raise HTTPError(404, 'Shift id not found:{}'.format(shift_id))
+
+            session.delete(shift)
+            session.commit()
+            return session.job(shift.job_id).to_dict(self.fields)
 
     @docstring_format(
         _format_opts(c.WORKED_STATUS_OPTS),
@@ -500,29 +590,28 @@ class JobLookup:
             status = int(status)
             assert c.WORKED_STATUS[status] is not None
         except Exception:
-            return {'error': 'Invalid status: {}'.format(status)}
+            raise HTTPError(400, 'Invalid status: {}'.format(status))
 
         try:
             rating = int(rating)
             assert c.RATINGS[rating] is not None
         except Exception:
-            return {'error': 'Invalid rating: {}'.format(rating)}
+            raise HTTPError(400, 'Invalid rating: {}'.format(rating))
 
         if rating in (c.RATED_BAD, c.RATED_GREAT) and not comment:
-            return {'error': 'You must leave a comment explaining why the '
-                    'staffer was rated as: {}'.format(c.RATINGS[rating])}
+            raise HTTPError(400, 'You must leave a comment explaining why the staffer was rated as: {}'.format(
+                c.RATINGS[rating]))
 
         with Session() as session:
-            try:
-                shift = session.shift(shift_id)
-                shift.worked = status
-                shift.rating = rating
-                shift.comment = comment
-                session.commit()
-            except Exception:
-                return {'error': 'Unexpected error setting status'}
-            else:
-                return session.job(shift.job_id).to_dict(self.fields)
+            shift = session.query(Shift).filter_by(id=shift_id).first()
+            if not shift:
+                raise HTTPError(404, 'Shift id not found:{}'.format(shift_id))
+
+            shift.worked = status
+            shift.rating = rating
+            shift.comment = comment
+            session.commit()
+            return session.job(shift.job_id).to_dict(self.fields)
 
 
 @all_api_auth(c.API_READ)
@@ -543,7 +632,9 @@ class DepartmentLookup:
         department ids call the "dept.list" method.
         """
         with Session() as session:
-            department = session.query(Department).get(department_id)
+            department = session.query(Department).filter_by(id=department_id).first()
+            if not department:
+                raise HTTPError(404, 'Department id not found: {}'.format(department_id))
             return department.to_dict({
                 'id': True,
                 'name': True,
@@ -552,6 +643,7 @@ class DepartmentLookup:
                 'is_shiftless': True,
                 'is_setup_approval_exempt': True,
                 'is_teardown_approval_exempt': True,
+                'max_consecutive_hours': 0,
                 'jobs': {
                     'id': True,
                     'type': True,
@@ -588,6 +680,9 @@ class ConfigLookup:
         'URL_BASE',
         'URL_ROOT',
         'PATH',
+        'BADGE_PRICE',
+        'BADGES_SOLD',
+        'REMAINING_BADGES',
     ]
 
     def info(self):
@@ -610,6 +705,8 @@ class ConfigLookup:
         """
         if field.upper() in self.fields:
             return getattr(c, field.upper())
+        else:
+            raise HTTPError(404, 'Config field not found: {}'.format(field))
 
 
 @all_api_auth(c.API_READ)
@@ -629,7 +726,7 @@ class BarcodeLookup:
             result = get_badge_num_from_barcode(barcode_value)
             badge_num = result['badge_num']
         except Exception as e:
-            return {'error': "Couldn't look up barcode value: " + str(e)}
+            raise HTTPError(500, "Couldn't look up barcode value: " + str(e))
 
         # Note: A decrypted barcode can yield a valid badge num,
         # but that badge num may not be assigned to an attendee.
@@ -640,7 +737,7 @@ class BarcodeLookup:
             if attendee:
                 return attendee.to_dict(fields)
             else:
-                return {'error': 'Valid barcode, but no attendee found with Badge #{}'.format(badge_num)}
+                raise HTTPError(404, 'Valid barcode, but no attendee found with Badge #{}'.format(badge_num))
 
     def lookup_badge_number_from_barcode(self, barcode_value):
         """
@@ -652,7 +749,7 @@ class BarcodeLookup:
             result = get_badge_num_from_barcode(barcode_value)
             return {'badge_num': result['badge_num']}
         except Exception as e:
-            return {'error': "Couldn't look up barcode value: " + str(e)}
+            raise HTTPError(500, "Couldn't look up barcode value: " + str(e))
 
 
 if c.API_ENABLED:
@@ -662,3 +759,4 @@ if c.API_ENABLED:
     register_jsonrpc(ConfigLookup(), 'config')
     register_jsonrpc(BarcodeLookup(), 'barcode')
     register_jsonrpc(GuestLookup(), 'guest')
+    register_jsonrpc(MivsLookup(), 'mivs')
